Check the correctness of data frame for separating

This function is called inside gr_separate(), but can be used explicitly inside your code.

Usage

gr_check_data(df)

Arguments

df

data.frame with four columns: date, discharge, temperature, precipitation, as required by gr_separate().

Value

stops execution if df contains the wrong number of columns, or the columns have the wrong types, or the data in columns is incorrect (e.g. discharge or precipitation are negative).

Examples

library(grwat)

# example Spas-Zagorye data is included with grwat package
data(spas)
print(spas)
#> # A tibble: 23,742 × 4
#>    Date           Q   Temp  Prec
#>    <date>     <dbl>  <dbl> <dbl>
#>  1 1956-01-01  5.18  -6.46 0.453
#>  2 1956-01-02  5.18 -11.4  0.825
#>  3 1956-01-03  5.44 -10.7  0.26 
#>  4 1956-01-04  5.44  -8.05 0.397
#>  5 1956-01-05  5.44 -11.7  0.102
#>  6 1956-01-06  5.58 -20.1  0.032
#>  7 1956-01-07  5.58 -16.6  0.024
#>  8 1956-01-08  5.36 -15.2  0.042
#>  9 1956-01-09  5.36 -15.9  0.053
#> 10 1956-01-10  5.31 -13.3  0.08 
#> # … with 23,732 more rows

gr_check_data(spas)
#> grwat: data frame is correct

# raw Spas-Zagorye data represents date components
# in columns and does not contain meteorologgical variables
path = system.file("extdata", "spas-zagorye.txt", 
                   package = "grwat")

hdata_raw = readr::read_delim(path, 
                       col_names = c('d', 'm', 'y', 'q'), 
                       col_types = 'iiid', delim = ' ')
#> Warning: One or more parsing issues, see `problems()` for details

print(hdata_raw)
#> # A tibble: 23,732 × 4
#>        d     m     y     q
#>    <int> <int> <int> <dbl>
#>  1     1     1  1956  5.18
#>  2     2     1  1956  5.18
#>  3     3     1  1956  5.44
#>  4     4     1  1956  5.44
#>  5     5     1  1956  5.44
#>  6     6     1  1956  5.58
#>  7     7     1  1956  5.58
#>  8     8     1  1956  5.36
#>  9     9     1  1956  5.36
#> 10    10     1  1956  5.31
#> # … with 23,722 more rows

gr_check_data(hdata_raw)
#> Error in gr_check_data(hdata_raw): grwat: the four columns of input data frame (date, discharge, temperature, precipitation) must be of Date, numeric, numeric, numeric data types