Test hydrograph changes

Use this function to test interannual changes or hydrograph separation variables returned by gr_summarize(). Pettitt test is used to detect the change year — i.e. the year which divides the time series into the statistically most differing samples. Student (Welch) and Fisher tests are used to estimate the significance of mean and variance differences of these samples. Theil-Sen test calculates the trend slope value. Mann-Kendall test is performed to reveal the significance of the trend.

Usage

gr_test_vars(df, ..., year = NULL, exclude = NULL)

Arguments

df

data.frame as produced by gr_summarize() function.

...

Names of the tested variables (quoted).

year

Integer value of year used to divide series in two samples compared by Student and Fisher tests. Defaults to NULL which means that the year is calculated automatically by Pettitt test.

exclude

Integer vector of years to be excluded from tests.

Value

list of testing results with following elements:

Element	Description

ptt	Pettitt tests for change year

mkt	Mann-Kendall test for trend significance

tst	Theil-Sen test for slope estimation

ts_fit	Theil-Sen linear model fit

tt	Student (Welch) test for significance of mean differences between two periods

ft	Fisher test for significance of variance differences between two periods

year	Integer value of year used to divide series in two samples compared by Student and Fisher tests

maxval	Maximum value for the variable along the full time series

fixed_year	Boolean TRUE or FALSE value indicating if the year was fixed

pvalues	p-values of all tests summarized as a single table for all variables

Details

Number of observations formally required for various tests: Pettitt > 0, Mann-Kendall > 2, Theil-Sen > 1, Student > 1, Fisher > 1.

Examples

# \donttest{
library(grwat)

data(spas) # example Spas-Zagorye data is included with grwat package

# separate
sep = gr_separate(spas, params = gr_get_params(reg = 'center'))
#> grwat: data frame is correct
#> grwat: parameters list and types are OK

# summarize
vars = gr_summarize(sep)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf

# test all variables
tests = gr_test_vars(vars)

# view Pettitt test for Qygr
tests$ptt$Qygr
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 869, p-value = 8.129e-08
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              23 
#> 

# view Fisher test for Q30s
tests$ft$Q30s
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.53023, num df = 21, denom df = 41, p-value = 0.1208
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.2597352 1.1887012
#> sample estimates:
#> ratio of variances 
#>          0.5302307 
#> 

# test only Qygr and Q30s using 1978 as fixed year and excluding 1988-1991 yrs
gr_test_vars(vars, Qygr, Q30s, year = 1978, exclude = 1988:1991)
#> $ptt
#> $ptt$Qygr
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 777, p-value = 1.372e-07
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              23 
#> 
#> 
#> $ptt$Q30s
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 661, p-value = 1.308e-05
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              23 
#> 
#> 
#> 
#> $mkt
#> $mkt$Qygr
#> 
#> 	Mann-Kendall trend test
#> 
#> data:  vl[vl_cmp]
#> z = 4.6495, n = 60, p-value = 3.327e-06
#> alternative hypothesis: true S is not equal to 0
#> sample estimates:
#>            S         varS          tau 
#> 7.300000e+02 2.458333e+04 4.124294e-01 
#> 
#> 
#> $mkt$Q30s
#> 
#> 	Mann-Kendall trend test
#> 
#> data:  vl[vl_cmp]
#> z = 3.272, n = 60, p-value = 0.001068
#> alternative hypothesis: true S is not equal to 0
#> sample estimates:
#>            S         varS          tau 
#> 5.140000e+02 2.458133e+04 2.905597e-01 
#> 
#> 
#> 
#> $tst
#> $tst$Qygr
#> 
#> 	Sen's slope
#> 
#> data:  df.theil[[2]][fltr]
#> z = 4.6495, n = 60, p-value = 3.327e-06
#> alternative hypothesis: true z is not equal to 0
#> 95 percent confidence interval:
#>  0.06251714 0.13936618
#> sample estimates:
#> Sen's slope 
#>  0.09723533 
#> 
#> 
#> $tst$Q30s
#> 
#> 	Sen's slope
#> 
#> data:  df.theil[[2]][fltr]
#> z = 3.272, n = 60, p-value = 0.001068
#> alternative hypothesis: true z is not equal to 0
#> 95 percent confidence interval:
#>  0.02823611 0.09316667
#> sample estimates:
#> Sen's slope 
#>  0.06039279 
#> 
#> 
#> 
#> $ts_fit
#> $ts_fit$Qygr
#> 
#> Call:
#> mblm::mblm(formula = eval(frml), dataframe = df.theil[fltr, ], 
#>     repeated = FALSE)
#> 
#> Coefficients:
#> (Intercept)        Year1  
#>  -161.12007      0.08728  
#> 
#> 
#> $ts_fit$Q30s
#> 
#> Call:
#> mblm::mblm(formula = eval(frml), dataframe = df.theil[fltr, ], 
#>     repeated = FALSE)
#> 
#> Coefficients:
#> (Intercept)        Year1  
#>   -101.1759       0.0549  
#> 
#> 
#> 
#> $tt
#> $tt$Qygr
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  d1 and d2
#> t = -9.026, df = 46.832, p-value = 8.113e-12
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -5.653447 -3.592487
#> sample estimates:
#> mean of x mean of y 
#>  9.141565 13.764532 
#> 
#> 
#> $tt$Q30s
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  d1 and d2
#> t = -6.7982, df = 47.426, p-value = 1.586e-08
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -4.104838 -2.230513
#> sample estimates:
#> mean of x mean of y 
#>  6.100841  9.268517 
#> 
#> 
#> 
#> $ft
#> $ft$Qygr
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.73431, num df = 20, denom df = 38, p-value = 0.4659
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.3517569 1.6868285
#> sample estimates:
#> ratio of variances 
#>          0.7343085 
#> 
#> 
#> $ft$Q30s
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.71161, num df = 20, denom df = 38, p-value = 0.4204
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.3408856 1.6346959
#> sample estimates:
#> ratio of variances 
#>          0.7116142 
#> 
#> 
#> 
#> $year
#> Qygr Q30s 
#> 1978 1978 
#> 
#> $maxval
#> $maxval$Qygr
#> [1] 20.05682
#> 
#> $maxval$Q30s
#> [1] 13.16333
#> 
#> 
#> $fixed_year
#> [1] TRUE
#> 
#> $pvalues
#>   N                                    Variable Change.Year   Trend      M1
#> 1 1 Mean annual groundwater ("baseflow") runoff        1978 0.08728 9.14156
#> 2 2       Minimum 30-day averaged summer runoff        1978 0.05490 6.10084
#>         M2 MeanRatio     sd1     sd2 sdRatio Mann.Kendall Pettitt Student
#> 1 13.76453      50.6 1.78279 2.08047    16.7      0.00000   0e+00       0
#> 2  9.26852      51.9 1.61105 1.90980    18.5      0.00107   1e-05       0
#>    Fisher
#> 1 0.46589
#> 2 0.42036
#> 
# }