Test hydrograph changes

Use this function to test interannual changes or hydrograph separation variables returned by gr_summarize(). Pettitt test is used to detect the change year — i.e. the year which divides the time series into the statistically most differing samples. Student (Welch) and Fisher tests are used to estimate the significance of mean and variance differences of these samples. Theil-Sen test calculates the trend slope value. Mann-Kendall test is performed to reveal the significance of the trend.

Usage

gr_test_vars(df, ..., year = NULL, exclude = NULL)

Arguments

df

data.frame as produced by gr_summarize() function.

...

Names of the tested variables (quoted).

year

Integer value of year used to divide series in two samples compared by Student and Fisher tests. Defaults to NULL which means that the year is calculated automatically by Pettitt test.

exclude

Integer vector of years to be excluded from tests.

Value

list of testing results

Details

Number of observations formally required for various tests: Pettitt > 0, Mann-Kendall > 2, Theil-Sen > 1, Student > 1, Fisher > 1.

Examples

library(grwat)

data(spas) # example Spas-Zagorye data is included with grwat package

# separate
sep = gr_separate(spas, params = gr_get_params(reg = 'Midplain'))
#> grwat: data frame is correct
#> grwat: parameters list and types are OK

# summarize
vars = gr_summarize(sep)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf

# test all variables
tests = gr_test_vars(vars)

# view Pettitt test for Qygr
tests$ptt$Qygr
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 822, p-value = 2.342e-07
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              22 
#> 

# view Fisher test for Q30s
tests$ft$Q30s
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.55655, num df = 20, denom df = 41, p-value = 0.1601
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.2703584 1.2704553
#> sample estimates:
#> ratio of variances 
#>          0.5565489 
#> 

# test only Qygr and Q30s using 1978 as fixed year and excluding 1988-1991 yrs
gr_test_vars(vars, Qygr, Q30s, year = 1978, exclude = 1988:1991)
#> $ptt
#> $ptt$Qygr
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 734, p-value = 3.797e-07
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              22 
#> 
#> 
#> $ptt$Q30s
#> 
#> 	Pettitt's test for single change-point detection
#> 
#> data:  vl[vl_cmp]
#> U* = 632, p-value = 2.079e-05
#> alternative hypothesis: two.sided
#> sample estimates:
#> probable change point at time K 
#>                              22 
#> 
#> 
#> 
#> $mkt
#> $mkt$Qygr
#> 
#> 	Mann-Kendall trend test
#> 
#> data:  vl[vl_cmp]
#> z = 4.2245, n = 59, p-value = 2.395e-05
#> alternative hypothesis: true S is not equal to 0
#> sample estimates:
#>            S         varS          tau 
#> 6.470000e+02 2.338367e+04 3.781414e-01 
#> 
#> 
#> $mkt$Q30s
#> 
#> 	Mann-Kendall trend test
#> 
#> data:  vl[vl_cmp]
#> z = 3.0998, n = 59, p-value = 0.001936
#> alternative hypothesis: true S is not equal to 0
#> sample estimates:
#>            S         varS          tau 
#> 4.750000e+02 2.338167e+04 2.777778e-01 
#> 
#> 
#> 
#> $tst
#> $tst$Qygr
#> 
#> 	Sen's slope
#> 
#> data:  df.theil[[2]][fltr]
#> z = 4.2245, n = 59, p-value = 2.395e-05
#> alternative hypothesis: true z is not equal to 0
#> 95 percent confidence interval:
#>  0.05523255 0.13548338
#> sample estimates:
#> Sen's slope 
#>  0.09043705 
#> 
#> 
#> $tst$Q30s
#> 
#> 	Sen's slope
#> 
#> data:  df.theil[[2]][fltr]
#> z = 3.0998, n = 59, p-value = 0.001936
#> alternative hypothesis: true z is not equal to 0
#> 95 percent confidence interval:
#>  0.02581481 0.09276190
#> sample estimates:
#> Sen's slope 
#>  0.05994444 
#> 
#> 
#> 
#> $ts_fit
#> $ts_fit$Qygr
#> 
#> Call:
#> mblm::mblm(formula = eval(frml), dataframe = df.theil[fltr, ], 
#>     repeated = FALSE)
#> 
#> Coefficients:
#> (Intercept)        Year1  
#>  -142.95366      0.07799  
#> 
#> 
#> $ts_fit$Q30s
#> 
#> Call:
#> mblm::mblm(formula = eval(frml), dataframe = df.theil[fltr, ], 
#>     repeated = FALSE)
#> 
#> Coefficients:
#> (Intercept)        Year1  
#>   -95.00467      0.05182  
#> 
#> 
#> 
#> $tt
#> $tt$Qygr
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  d1 and d2
#> t = -8.5734, df = 41.956, p-value = 9.076e-11
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -5.468779 -3.384706
#> sample estimates:
#> mean of x mean of y 
#>  8.950695 13.377437 
#> 
#> 
#> $tt$Q30s
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  d1 and d2
#> t = -6.6349, df = 43.712, p-value = 4.045e-08
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -4.147975 -2.214893
#> sample estimates:
#> mean of x mean of y 
#>  6.087083  9.268517 
#> 
#> 
#> 
#> $ft
#> $ft$Qygr
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.8213, num df = 19, denom df = 38, p-value = 0.6593
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.3899576 1.9240464
#> sample estimates:
#> ratio of variances 
#>           0.821299 
#> 
#> 
#> $ft$Q30s
#> 
#> 	F test to compare two variances
#> 
#> data:  d1 and d2
#> F = 0.74792, num df = 19, denom df = 38, p-value = 0.5045
#> alternative hypothesis: true ratio of variances is not equal to 1
#> 95 percent confidence interval:
#>  0.355117 1.752143
#> sample estimates:
#> ratio of variances 
#>          0.7479204 
#> 
#> 
#> 
#> $year
#> Qygr Q30s 
#> 1978 1978 
#> 
#> $maxval
#> $maxval$Qygr
#> [1] 19.52822
#> 
#> $maxval$Q30s
#> [1] 13.16333
#> 
#> 
#> $fixed_year
#> [1] TRUE
#> 
#> $pvalues
#>   N                                    Variable Change.Year   Trend      M1
#> 1 1 Mean annual groundwater ("baseflow") runoff        1978 0.07799 8.95069
#> 2 2       Minimum 30-day averaged summer runoff        1978 0.05182 6.08708
#>         M2 MeanRatio     sd1     sd2 sdRatio Mann.Kendall Pettitt Student
#> 1 13.37744      49.5 1.81175 1.99916    10.3      0.00002   0e+00       0
#> 2  9.26852      52.3 1.65164 1.90980    15.6      0.00194   2e-05       0
#>    Fisher
#> 1 0.65931
#> 2 0.50455
#>